requirements.txt
numpy
scipy
matplotlib
torch==0.4.1
torchvision==0.2.1
nltk
Pillow
tqdm
lmdb
tensorflow
tensorboardX
